migrating to v0.15 of the jsr 170 spec
git-svn-id: https://svn.apache.org/repos/asf/incubator/jackrabbit/trunk@55516 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/jackrabbit/core/observation/ObservationManagerFactory.java
@@ -268,7 +268,7 @@ public final class ObservationManagerFactory implements Runnable {
          * @see ObservationManager#addEventListener
          */
         public void addEventListener(EventListener listener,
-                                     long eventTypes,
+                                     int eventTypes,
                                      String absPath,
                                      boolean isDeep,
                                      String[] uuid,
